adding sandboxing refactored
